Match and link grpstadium (#3092)
* Widen gm_80160968 parameter to u32

The DOL never extends gm_80160854's result before passing it to gm_80160968 (ifmagnify, grpstadium call sites), while gm_80160968's own body masks the index to a byte (rlwinm 2,22,29). A u32 parameter with a u8 cast in the body reproduces both: callers pass the u32 result unconverted and the callee's masked index is unchanged. gm_80160854 and gm_80160968 both stay 100%; removes a spurious caller-side clrlwi in grpstadium, ifmagnify, and gmresultplayer.

* Improve grStadium_801D3BBC match

Move grPs_803B7F8C's definition below the function behind an extern declaration so its address materializes at first use instead of hoisting into the prologue, and copy the four alpha presets with a single 16-byte struct assignment as the paired lwz/stw stream shows. 91.9% -> 99.8%; the remaining rows are one callee-saved color pair on the assert-string base web.

* Improve grStadium_801D4548 match

Remove two dozen unused m2c locals, correct the case-3 counter test to a post-increment (the DOL compares the pre-increment value), split the case-5 increment from its bound check so the counter reloads, accumulate the case-4 scale in place so it claims f29 across the setter's assert call, keep the jobj copy web alive for the nested SetMtxDirtySub call, and pad the frame to the retail layout. 98.3% -> 99.6%; remaining rows are stack-slot ordering for the case-0 array and a few scratch-register rotations.

* Match grpstadium data sections

The named grPs_804DAFxx floats were decomp artifacts: the DOL's .sdata2 is the anonymous literal pool, deduplicated by value at first textual use, and defining the floats as named objects both duplicated pool entries (a visible const float definition const-folds to a fresh literal at later use sites) and misplaced them. Drop the extern declarations and definitions and spell the values as literals; .sdata2 is now byte-identical. Shrink grPs_8049F040 to the u8[0x100] buffer the object actually reserves. Remaining section deltas are tail alignment the linker fills.

* Match grStadium_801D4548 frame layout

Stack slots allocate in reverse declaration order, so the dead pad and the two homed Ground pointers must be declared after case 0's sp60 array to sit below it in the frame: declare temp_r6 and temp_r3_9 in the case blocks that use them and move the pad into case 6. The sp60 array and conversion staging now land at retail offsets.

* Match grStadium_801D3BBC

The hud color call takes its state and slot arguments as nested calls rather than u8 locals: as named locals their webs rank in the user band and select too late, leaving the assert-string base and constant-1 webs to claim r27; as call-argument temporaries they claim r25/r26 early (MWCC evaluates arguments right to left, preserving the call order) and the base web recycles r25 as retail does. Dropping the two now-unused locals also restores the colors union's stack offset.

* Improve grStadium_801D4548 match

A shadow use for temp_r5_2 rebands its value web so the halved bound's web selects first and the case-5 else scratch registers land as retail. 99.59% -> 99.66%.

* Reformat gm_1601.dox with pinned clang-format

The comment alignment differed between clang-format 22.1.8 (local) and the 22.1.5 the CI lint hook pins.

* Improve grStadium_801D4548 match

The pokemon pick loop computes its random index in the body and does the sp60 lookup inside the loop condition, so the xE2 comparison load is emitted before the index shift; this reproduces the retail scratch-register order in the do-while. 99.66% -> 99.69%.

* Improve grStadium_801D4548 match

Declaring var_f30 ahead of the other float locals ranks its register web between the two case-4 scale temporaries, so it selects a callee-saved color before f29 is claimed and lands in f30 as the retail does instead of reusing f29. 99.69% -> 99.71%; only the case-4 GET_JOBJ load/home ordering remains.

* Match grStadium_801D4548

* Match and link grpstadium

* Improve ifStatus_802F61FC and ifMagnify_802FB8C0 matching

Recovers the two regressions from widening gm_80160968 to u32 and goes further: ifStatus_802F61FC 96.64% -> 98.31%, ifMagnify_802FB8C0 98.52% -> 99.95%.

+++ src/melee/gm/gm_1601.c
@@ -692,9 +692,9 @@ u32 gm_80160854(u8 slot, u8 team, u8 is_teams, u8 slot_type)
     return 0;
 }
 
-GXColor gm_80160968(u8 arg0)
+GXColor gm_80160968(u32 arg0)
 {
-    return lbl_803B7864[arg0];
+    return lbl_803B7864[(u8) arg0];
 }
 
 const char* gm_80160980_noinline(u8 ckind);
